requirements.txt
ollama
openai
cohere
google-generativeai
groq
transformers